+++ packages/giselle-engine/src/react/flow/hooks/index.ts
@@ -3,5 +3,4 @@ export * from "./use-copy-node";
 export * from "./use-connection";
 export * from "./use-node-update";
 export * from "./use-properties-panel";
-export * from "./use-view";
 export * from "./use-workflow-designer";
