+++ packages/giselle-engine/src/react/flow/hooks/use-properties-panel.ts
