+++ packages/giselle-engine/src/react/flow/utils/is-supported-connection.ts
