+++ packages/giselle-engine/src/utils/node-default-name.ts
@@ -8,7 +8,7 @@ import {
 	isVectorStoreNode,
 } from "@giselle-sdk/data-type";
 import type { ActionProvider, TriggerProvider } from "@giselle-sdk/flow";
-import type { VectorStoreSourceProvider } from "@giselle-sdk/giselle-engine";
+import type { VectorStoreSourceProvider } from "../core/vector-store";
 
 export const triggerProviderLabel: Record<TriggerProvider, string> = {
 	github: "GitHub",
