+++ packages/giselle-sdk/src/react/index.ts
@@ -1,3 +1,2 @@
 export * from "@giselle-sdk/workspace/react";
 export { useGiselleEngine } from "@giselle-sdk/giselle-engine/react";
-export { ViewState } from "@giselle-sdk/workflow-designer/react";
