--- packages/node-utils/package.json
@@ -1,34 +1,0 @@
-{
-	"name": "@giselle-sdk/node-utils",
-	"version": "0.0.0",
-	"private": true,
-	"main": "./dist/index.js",
-	"module": "./dist/index.mjs",
-	"types": "./dist/index.d.ts",
-	"scripts": {
-		"build": "tsup",
-		"check-types": "tsc --noEmit",
-		"clean": "rm -rf .turbo dist node_modules",
-		"format": "biome check --write ."
-	},
-	"exports": {
-		"./package.json": "./package.json",
-		".": {
-			"types": "./dist/index.d.ts",
-			"import": "./dist/index.mjs",
-			"require": "./dist/index.js"
-		}
-	},
-	"devDependencies": {
-		"@giselle/giselle-sdk-tsconfig": "workspace:*",
-		"tsup": "catalog:"
-	},
-	"dependencies": {
-		"@giselle-sdk/data-type": "workspace:*",
-		"@giselle-sdk/flow": "workspace:*",
-		"@giselle-sdk/giselle-engine": "workspace:^",
-		"@giselle-sdk/language-model": "workspace:*",
-		"@giselle-sdk/text-editor-utils": "workspace:*",
-		"@tiptap/react": "catalog:"
-	}
-}
