--- packages/node-utils/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./default-name";
-export * from "./node-factories";
