--- packages/node-utils/tsup.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig([
-	{
-		entry: ["src/index.ts"],
-		format: ["cjs", "esm"],
-		dts: true,
-		sourcemap: true,
-	},
-]);
