--- packages/node-utils/turbo.json
@@ -1,8 +1,0 @@
-{
-	"extends": ["//"],
-	"tasks": {
-		"build": {
-			"outputs": ["**/dist/**"]
-		}
-	}
-}
