+++ packages/text-editor-utils/src/index.ts
@@ -2,3 +2,4 @@ export * from "./extensions";
 export * from "./is-json-content";
 export * from "./json-content-to-text";
 export * from "./node-references";
+export type { JSONContent } from "@tiptap/react";
