--- packages/workflow-designer/package.json
@@ -1,42 +1,0 @@
-{
-	"name": "@giselle-sdk/workflow-designer",
-	"version": "0.0.0",
-	"private": true,
-	"main": "./dist/index.js",
-	"module": "./dist/index.mjs",
-	"types": "./dist/index.d.ts",
-	"scripts": {
-		"build": "tsup",
-		"check-types": "tsc --noEmit",
-		"clean": "rm -rf .turbo dist node_modules",
-		"format": "biome check --write .",
-		"test": "vitest run"
-	},
-	"files": ["dist/**/*"],
-	"exports": {
-		"./package.json": "./package.json",
-		".": {
-			"types": "./dist/index.d.ts",
-			"import": "./dist/index.mjs",
-			"require": "./dist/index.js",
-			"default": "./dist/index.js"
-		},
-		"./dist/style.css": "./dist/style.css"
-	},
-	"devDependencies": {
-		"@giselle/giselle-sdk-tsconfig": "workspace:*",
-		"@types/react": "catalog:",
-		"tsup": "catalog:",
-		"vitest": "catalog:"
-	},
-	"dependencies": {
-		"@giselle-sdk/data-type": "workspace:*",
-		"@giselle-sdk/giselle-engine": "workspace:*",
-		"@giselle-sdk/language-model": "workspace:^",
-		"@giselle-sdk/node-utils": "workspace:*",
-		"@giselle-sdk/text-editor-utils": "workspace:*",
-		"@giselle-sdk/workflow-utils": "workspace:*",
-		"react": "catalog:",
-		"zod": "catalog:"
-	}
-}
