--- packages/workflow-designer/tsconfig.json
@@ -1,10 +1,0 @@
-{
-	"extends": "@giselle/giselle-sdk-tsconfig/react-library.json",
-	"include": ["."],
-	"exclude": ["node_modules"],
-	"compilerOptions": {
-		"target": "ES2018",
-		"module": "ESNext",
-		"moduleResolution": "bundler"
-	}
-}
