--- packages/workflow-designer/tsup.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig({
-	entry: ["src/index.ts"],
-	format: ["cjs", "esm"],
-	dts: true,
-	sourcemap: true,
-});
