+++ packages/workspace/src/react/hooks.ts
@@ -2,5 +2,5 @@ export {
 	useNodeGenerations,
 	useUsageLimits,
 	useVectorStore,
+	useWorkflowDesigner,
 } from "@giselle-sdk/giselle-engine/react";
-export { useWorkflowDesigner } from "@giselle-sdk/workflow-designer/react";
