+++ packages/workspace/src/react/index.ts
@@ -1,6 +1,5 @@
 "use client";
 
-export type { ViewState } from "@giselle-sdk/workflow-designer/react";
 export * from "./feature-flag";
 export * from "./hooks";
 export * from "./workspace";
